azure-functions/requirements.txt
# Azure Functions requirements
azure-functions
azure-storage-blob
requests
numpy
